requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
scikit-learn
numpy
faiss-cpu
lxml
